Allow --use-application-binary to use app-bundles

+++ packages/flutter_tools/lib/src/application_package.dart
@@ -145,29 +145,36 @@ bool _isBundleDirectory(FileSystemEntity entity) =>
 abstract class IOSApp extends ApplicationPackage {
   IOSApp({@required String projectBundleId}) : super(id: projectBundleId);
 
-  /// Creates a new IOSApp from an existing IPA.
-  factory IOSApp.fromIpa(String applicationBinary) {
+  /// Creates a new IOSApp from an existing app bundle or IPA.
+  factory IOSApp.fromPrebuiltApp(String applicationBinary) {
     Directory bundleDir;
-    try {
-      final Directory tempDir = fs.systemTempDirectory.createTempSync('flutter_app_');
-      addShutdownHook(() async {
-        await tempDir.delete(recursive: true);
-      }, ShutdownStage.STILL_RECORDING);
-      os.unzip(fs.file(applicationBinary), tempDir);
-      final Directory payloadDir = fs.directory(fs.path.join(tempDir.path, 'Payload'));
-      bundleDir = payloadDir.listSync().singleWhere(_isBundleDirectory);
-    } on StateError catch (e, stackTrace) {
-      printError('Invalid prebuilt iOS binary: ${e.toString()}', stackTrace: stackTrace);
-      return null;
+    if (fs.isDirectorySync(applicationBinary)) {
+      // Looks like an app bundle.
+      bundleDir = fs.directory(applicationBinary);
+    } else {
+      // Looks like an ipa.
+      try {
+        final Directory tempDir = fs.systemTempDirectory.createTempSync(
+            'flutter_app_');
+        addShutdownHook(() async {
+          await tempDir.delete(recursive: true);
+        }, ShutdownStage.STILL_RECORDING);
+        os.unzip(fs.file(applicationBinary), tempDir);
+        final Directory payloadDir = fs.directory(
+            fs.path.join(tempDir.path, 'Payload'));
+        bundleDir = payloadDir.listSync().singleWhere(_isBundleDirectory);
+      } on StateError catch (e, stackTrace) {
+        printError('Invalid prebuilt iOS binary: ${e.toString()}',
+            stackTrace: stackTrace);
+        return null;
+      }
     }
-
     final String plistPath = fs.path.join(bundleDir.path, 'Info.plist');
     final String id = plist.getValueFromFile(plistPath, plist.kCFBundleIdentifierKey);
     if (id == null)
       return null;
 
     return new PrebuiltIOSApp(
-      ipaPath: applicationBinary,
       bundleDir: bundleDir,
       bundleName: fs.path.basename(bundleDir.path),
       projectBundleId: id,
@@ -237,12 +244,10 @@ class BuildableIOSApp extends IOSApp {
 }
 
 class PrebuiltIOSApp extends IOSApp {
-  final String ipaPath;
   final Directory bundleDir;
   final String bundleName;
 
   PrebuiltIOSApp({
-    this.ipaPath,
     this.bundleDir,
     this.bundleName,
     @required String projectBundleId,
@@ -274,7 +279,7 @@ Future<ApplicationPackage> getApplicationPackageForPlatform(TargetPlatform platf
     case TargetPlatform.ios:
       return applicationBinary == null
           ? new IOSApp.fromCurrentDirectory()
-          : new IOSApp.fromIpa(applicationBinary);
+          : new IOSApp.fromPrebuiltApp(applicationBinary);
     case TargetPlatform.tester:
       return new FlutterTesterApp.fromCurrentDirectory();
     case TargetPlatform.darwin_x64:
